fixed broken free and dereferenceing of pointer to of buf message

+++ src/lib/testclients/FakeClientSockets/FakeSocketInterface.h
@@ -15,7 +15,7 @@ public:
 
     virtual bool isDisconnected() =0;
 
-    virtual ssize_t send(const uint8_t *buf, size_t len)=0;
+    virtual ssize_t send(const uint8_t *buf, uint8_t len)=0;
 
     virtual void connect(device_address *address)=0;
